+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - GCGI-1459: Update report date for PWGS assay from yyyy/mm/dd to date report was generated
 - GCGI-1458: Change geneticist sign-off date from date report was generated to yyyy-mm-dd
 - GCGI-1465: Fix for overzealous date format check from implementation of GCGI-1458
+- GCGI-1326: Generate whizbam links to view fusions in split-view
 
 ## 1.7.6: 2024-10-22
 - GCGI-1156: Merge long-running documentation branch to main